backend/requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
python-dotenv==1.0.0
aiogram==2.25.1
geopy==2.3.0
requests==2.31.0